# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: ProfilFinder.at API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /profiles:
    get:
      operationId: getProfiles
      tags: [profiles]
      summary: Get public profiles with filters
      parameters:
        - in: query
          name: search
          schema:
            type: string
        - in: query
          name: industry
          schema:
            type: string
        - in: query
          name: federalState
          schema:
            type: string
        - in: query
          name: availability
          schema:
            type: string
        - in: query
          name: profession
          schema:
            type: string
        - in: query
          name: page
          schema:
            type: integer
        - in: query
          name: limit
          schema:
            type: integer
      responses:
        "200":
          description: List of public profiles
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProfileListResponse"
    post:
      operationId: createProfile
      tags: [profiles]
      summary: Create a profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProfileInput"
      responses:
        "201":
          description: Profile created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicProfile"
  /profiles/{profileNumber}:
    get:
      operationId: getProfileByNumber
      tags: [profiles]
      summary: Get a public profile by profile number
      parameters:
        - in: path
          name: profileNumber
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Public profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicProfile"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /profiles/{id}/update:
    patch:
      operationId: updateProfile
      tags: [profiles]
      summary: Update a profile
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProfileUpdate"
      responses:
        "200":
          description: Profile updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicProfile"
  /profiles/{id}/delete:
    delete:
      operationId: deleteProfile
      tags: [profiles]
      summary: Delete a profile
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /my-profile:
    get:
      operationId: getMyProfile
      tags: [profiles]
      summary: Get the current user's profile
      responses:
        "200":
          description: My profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicProfile"
        "404":
          description: No profile yet
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
    MessageResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string
...
    PublicProfile:
      type: object
      required: [id, profileNumber, jobTitle, industry, federalState, availability, isAvailable, createdAt]
      properties:
        id:
          type: integer
        profileNumber:
          type: string
        publicName:
          type: ["string", "null"]
        jobTitle:
          type: string
        profession:
          type: ["string", "null"]
        industry:
          type: string
        experience:
          type: ["string", "null"]
        education:
          type: ["string", "null"]
        skills:
          type: ["string", "null"]
        languages:
          type: ["string", "null"]
        desiredJob:
          type: ["string", "null"]
        federalState:
          type: string
        city:
          type: ["string", "null"]
        availability:
          type: string
        employmentType:
          type: ["string", "null"]
        driverLicense:
          type: boolean
        description:
          type: ["string", "null"]
        cvText:
          type: ["string", "null"]
        profileImage:
          type: ["string", "null"]
        isAvailable:
          type: boolean
        isPublic:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: ["string", "null"]
          format: date-time
    ProfileListResponse:
      type: object
      required: [profiles, total, page, limit]
      properties:
        profiles:
          type: array
          items:
            $ref: "#/components/schemas/PublicProfile"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
    ProfileInput:
      type: object
      required: [jobTitle, industry, federalState, availability, description]
      properties:
        firstName:
          type: string
        lastName:
          type: string
        publicName:
          type: string
        jobTitle:
          type: string
        profession:
          type: string
        industry:
          type: string
        experience:
          type: string
        education:
          type: string
        skills:
          type: string
        languages:
          type: string
        desiredJob:
          type: string
        federalState:
          type: string
        city:
          type: string
        availability:
          type: string
        employmentType:
          type: string
        driverLicense:
          type: boolean
        description:
          type: string
        cvText:
          type: string
        profileImage:
          type: string
        isAvailable:
          type: boolean
        isPublic:
          type: boolean
        phoneNumber:
          type: string
        contactEmail:
          type: string
        address:
          type: string
    ProfileUpdate:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        publicName:
          type: string
        jobTitle:
          type: string
        profession:
          type: string
        industry:
          type: string
        experience:
          type: string
        education:
          type: string
        skills:
          type: string
        languages:
          type: string
        desiredJob:
          type: string
        federalState:
          type: string
        city:
          type: string
        availability:
          type: string
        employmentType:
          type: string
        driverLicense:
          type: boolean
        description:
          type: string
        cvText:
          type: string
        profileImage:
          type: string
        isAvailable:
          type: boolean
        isPublic:
          type: boolean
        phoneNumber:
          type: string
        contactEmail:
          type: string
        address:
          type: string
...