# lib/api-zod/src/generated/types/profileListResponse.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * ProfilFinder.at API
 * OpenAPI spec version: 0.1.0
 */
import type { PublicProfile } from './publicProfile';

export interface ProfileListResponse {
  profiles: PublicProfile[];
  total: number;
  page: number;
  limit: number;
}