# lib/api-zod/src/generated/types/profileUpdate.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * ProfilFinder.at API
 * OpenAPI spec version: 0.1.0
 */

export interface ProfileUpdate {
  firstName?: string;
  lastName?: string;
  publicName?: string;
  jobTitle?: string;
  profession?: string;
  industry?: string;
  experience?: string;
  education?: string;
  skills?: string;
  languages?: string;
  desiredJob?: string;
  federalState?: string;
  city?: string;
  availability?: string;
  employmentType?: string;
  driverLicense?: boolean;
  description?: string;
  cvText?: string;
  profileImage?: string;
  isAvailable?: boolean;
  isPublic?: boolean;
  phoneNumber?: string;
  contactEmail?: string;
  address?: string;
}